hackthon/requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
joblib>=1.5.2
numpy>=2.3.3
openai>=1.107.2
pandas>=2.3.2
python-dotenv>=1.1.1
scikit-learn>=1.7.2
spacy>=3.8.7
werkzeug>=3.1.3
google-genai>=0.8.2